S3 lookup
Quick hack to allow the user pass a bucket and an optional prefix so the sample
`MetadataCreate` message is populated with the contents found. There is a hard
limit on 500 items, in realith the object storage returns up to 1000 matches
and more can be fetched paging the results, but I'm not doing that to keep it
simple.